Extract presentational cores from tool approval and personal auth components
Extract ToolValidationCard and PersonalAuthenticationCard out of MCPToolValidationRequired and
MCPServerPersonalAuthenticationRequired so the cards can be reused outside of conversations
(upcoming sandbox function tool approval in the viz wrapper). The conversation components become
thin wrappers keeping the conversation-specific parts (useValidateAction/useResolveAuthentication,
BlockedActionsProvider context, always-allow cascade). No behavior change.
