+++ front/components/assistant/conversation/ToolValidationDetails.tsx
@@ -5,7 +5,7 @@ import {
 import { PodMembersUpdateValidationDetails } from "@app/components/assistant/conversation/tool_validation/PodMembersUpdateValidationDetails";
 import { PodTasksCreateValidationDetails } from "@app/components/assistant/conversation/tool_validation/PodTasksCreateValidationDetails";
 import { PodTasksUpdateValidationDetails } from "@app/components/assistant/conversation/tool_validation/PodTasksUpdateValidationDetails";
-import type { BlockedToolExecution } from "@app/lib/actions/mcp";
+import type { ToolExecutionBase } from "@app/lib/actions/mcp_internal_actions/events";
 import { ASHBY_SERVER_NAME } from "@app/lib/actions/mcp_internal_actions/constants";
 import {
   CREATE_REFERRAL_TOOL_NAME,
@@ -80,7 +80,9 @@ interface DisplayableInput {
 }
 
 interface ToolValidationDetailsProps {
-  blockedAction: BlockedToolExecution;
+  // Only the display fields are needed, so both agent-loop and sandbox-function blocked tool
+  // executions can be rendered.
+  blockedAction: Pick<ToolExecutionBase, "inputs" | "metadata">;
   user: UserType;
   owner: LightWorkspaceType;
   conversationId?: string | null;
